додав кнопку для профіля юзера + невеличкий рефактор навігації
