+++ src/components/AuthNav/AuthNav.jsx
@@ -1,5 +1,6 @@
 import { NavLink } from 'react-router-dom';
 import styles from './AuthNav.module.scss';
+
 export default function AuthNav({ toggleMenu }) {
   return (
     <div className={styles.authNavWrap}>
