+++ src/components/AuthNav/AuthNav.module.scss
@@ -7,10 +7,12 @@
   @media screen and (min-width: 768px) {
     margin-bottom: 0;
     margin-right: 20px;
+    margin-left: 246px;
   }
   @media screen and (min-width: 1280px) {
     margin-bottom: 0;
-    margin-right: 0;
+    margin-right: 0px;
+    margin-left: 360px;
   }
 }
 .authNavLink {
