+++ src/components/Header/Header.module.scss
@@ -14,28 +14,3 @@
   @media screen and (min-width: 1280px) {
   }
 }
-
-.mainNav {
-  @media screen and (max-width: 1279px) {
-    background-color: $backgroundColor;
-    position: absolute;
-    top: 100px;
-    right: 0;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    transition: 1s;
-    transform: translateY(-100vh);
-    justify-content: flex-start;
-  }
-
-  @media screen and (min-width: 1280px) {
-    display: flex;
-    align-items: center;
-  }
-}
-.responsiveNav {
-  transform: none;
-  height: 100vh;
-  z-index: 100;
-}
