+++ src/components/Logo/Logo.module.scss
@@ -17,7 +17,7 @@
     font-size: 32px;
     line-height: 48px;
 
-    margin-right: 246px;
+    margin-right: 0px;
   }
   @media screen and (min-width: 1280px) {
     margin-right: 80px;
