+++ src/components/Nav/Nav.module.scss
@@ -9,7 +9,6 @@
 
   @media screen and (min-width: 1280px) {
     display: block;
-    margin-right: 360px;
   }
 }
